audio: fix gst_audio_format_from_caps() for 8-bit formats
The endianness field is not present in 8-bit audio caps, so don't
fail because of that when parsing caps into an GstAudioInfo structure.
And if it is present, it should be ignored, because the format
table we match against has a 0 value for endianness, so we need
to look for 0 as well.

https://bugzilla.gnome.org/show_bug.cgi?id=693166